+++ packages/cli/CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Added a check to `moon docker scaffold` that ensures that `.moon/cache` is ignored in a root
   `.dockerignore` file. This helps to avoid interoperability issues.
+- Added more logs to `moon docker` commands to help uncover future issues.
 
 ## 1.25.1